Refactor Operand class
